Data Modeling Cassandra/requirements.txt
cassandra-driver==3.24.0
boto3==1.15.6
ipython-sql==0.4.0